# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student manager",
        "description": "Documentation for student manager",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/classrooms": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get a list of classrooms",
                "description": "Returns a list of classrooms",
                "operationId": "2f72f5470de53ee3c1b98260a19b95ac",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Create a new classroom",
                "description": "Creates a new classroom",
                "operationId": "8e2ce7d8f98adf8e177ee807f293c2f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Scalable Systems Architecture"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Learn how to design and implement robust systems that handle high traffic and scale efficiently."
                                    },
                                    "type": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/classrooms/{id}": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get classroom by id",
                "description": "Return a classroom by id",
                "operationId": "d1dffef915059d675ed37f532aef10f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Classroom not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update an existing classroom",
                "description": "Updates an existing classroom",
                "operationId": "a4c88d898890ebf48d548695b3eb0718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Scalable Systems Architecture"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Learn how to design and implement robust systems that handle high traffic and scale efficiently."
                                    },
                                    "type": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "404": {
                        "description": "Classroom not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Delete a classroom",
                "description": "Delete a classroom",
                "operationId": "6a60bd3281c5568275ae795e00ba0e0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Classroom not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Autenticação usando Bearer Token. Formato: 'Bearer <token>'",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}